Added frontend repository in the context
The original frontend design has been added as a submodule so that it
can be used within context.

This is an initial approach as other methods can be adopted to obtain
sources and compile them within the image.

It all depends on the context of use: Development, Testing, Production.

In the case of development the ideal is that the submodule is mounted as
an external folder and thus internally accessed a folder to compile and
test the app.

In the case of testing, sources can be obtained directly in context or
downloaded via git and thus tested regularly each time it becomes
apparent that new commits can be checked from time to time.

In the case of production environment, the ideal approach is to verify
with the adopted container server the best approach.